Simplify the KisPainter API
Remove the overloaded bitBlt functions with opacity and composite
op arguments; also remove the bltSelection functions. This leads
to some reduction in code duplication and an api that's more in
line with Qt4. (We started this before, but never completely finished
it).

Also bring the rest of Krita in line with the new api.

git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/koffice@970880 283d02a7-25f6-0310-bc7c-ecb5cbfe19da