A plugin to set the background pattern of an image
Make it possible to select a pattern that will be used as
the background for an image.

TODO: set the pattern on the image, make it possible to
select solid colors using this plugin (by having categories
in the icon view), enable the add/delete/reset buttons,
make it possible to create patters from a selection or
the current layer.

git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/koffice@971290 283d02a7-25f6-0310-bc7c-ecb5cbfe19da